# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Academic World Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mongo/publication-details:
    get:
      operationId: mongoPublicationDetails
      tags: [mongo]
      summary: Full-text publication search
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicationDetail"
  /mongo/faculty-keyword-scores:
    get:
      operationId: mongoFacultyKeywordScores
      tags: [mongo]
      summary: Faculty ranked by score for a keyword
      parameters:
        - in: query
          name: keyword
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FacultyScore"
  /mongo/keyword-stats:
    get:
      operationId: mongoKeywordStats
      tags: [mongo]
      summary: Aggregate stats for a keyword
      parameters:
        - in: query
          name: keyword
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeywordStats"
  /mongo/top-cited:
    get:
      operationId: mongoTopCited
      tags: [mongo]
      summary: Top cited publications
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicationDetail"
  /mongo/venue-analysis:
    get:
      operationId: mongoVenueAnalysis
      tags: [mongo]
      summary: Publication venue citation impact analysis (facet pipeline)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 15
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VenueStats"
...
components:
  schemas:
...
    PublicationDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        year:
          type: integer
        numCitations:
          type: integer
        venue:
          type: string
        abstract:
          type: string
        keywords:
          type: array
          items:
            type: string
      required: [id, title]
    FacultyScore:
      type: object
      properties:
        name:
          type: string
        university:
          type: string
        score:
          type: number
      required: [name, score]
...
    KeywordStats:
      type: object
      properties:
        keyword:
          type: string
        totalPublications:
          type: integer
        totalCitations:
          type: integer
        avgCitations:
          type: number
        topVenue:
          type: string
      required: [keyword, totalPublications, totalCitations, avgCitations]
    VenueStats:
      type: object
      properties:
        venue:
          type: string
        totalPublications:
          type: integer
        totalCitations:
          type: integer
        avgCitations:
          type: number
      required: [venue, totalPublications, totalCitations, avgCitations]
...